Export as plain text with page break symbol

Export as plain text with page break symbol

Usage

export_as_txt(
  lst,
  file = NULL,
  page_type = NULL,
  landscape = FALSE,
  pg_width = page_dim(page_type)[if (landscape) 2 else 1],
  pg_height = page_dim(page_type)[if (landscape) 1 else 2],
  font_family = "Courier",
  font_size = 8,
  paginate = .need_pag(page_type, pg_width, pg_height, lpp, cpp),
  cpp = NULL,
  lpp = NULL,
  ...,
  page_break = "\\s\\n",
  hsep = default_hsep(),
  indent_size = 2,
  tf_wrap = paginate,
  max_width = cpp,
  colwidths = propose_column_widths(matrix_form(lst, TRUE))
)

Arguments

lst

Listing object

file

character(1). File to write.

page_type

character(1). Name of a page type. See page_types. Ignored when pg_width and pg_height are set directly.

landscape

logical(1). Should the dimensions of page_type be inverted for landscape? Defaults to FALSE, ignored when pg_width and pg_height are set directly.

pg_width

numeric(1). Page width in inches.

pg_height

numeric(1). Page height in inches.

font_family

character(1). Name of a font family. An error will be thrown if the family named is not monospaced. Defaults to Courier.

font_size

numeric(1). Font size, defaults to 8.

paginate

logical(1). Should lst be paginated before writing the file. Defaults to TRUE if any sort of page dimension is specified.

cpp

numeric(1) or NULL. Width (in characters) of the pages for horizontal pagination. NULL (the default) indicates no horizontal pagination should be done.

lpp

numeric(1) or NULL. Number of row lines (not counting titles and footers) to have per page. Standard is 70 while NULL disables vertical pagination.

...

Passed directly to paginate_listing

page_break

character(1). Page break symbol (defaults to outputting "\s").

hsep

character(1). Characters to repeat to create header/body separator line.

indent_size

numeric(1). Indent size in characters. Ignored when x is already a MatrixPrintForm object in favor of information there.

tf_wrap

logical(1). Should the texts for title, subtitle, and footnotes be wrapped?

max_width

integer(1), character(1) or NULL. Width that title and footer (including footnotes) materials should be word-wrapped to. If NULL, it is set to the current print width of the session (getOption("width")). If set to "auto", the width of the table (plus any table inset) is used. Ignored completely if tf_wrap is FALSE.

colwidths

numeric vector. Column widths (in characters) for use with vertical pagination.

Value

file (this function is called for the side effect of writing the file.

Note

When specified, font_size is used only to determine pagination based on page dimensions. The written file is populated in raw ASCII text, which does not have the concept of font size.

Examples

dat <- ex_adae
lsting <- as_listing(dat[1:25, ], key_cols = c("USUBJID", "AESOC")) %>%
  add_listing_col("AETOXGR") %>%
  add_listing_col("BMRKR1", format = "xx.x") %>%
  add_listing_col("AESER / AREL", fun = function(df) paste(df$AESER, df$AREL, sep = " / "))
#> sorting incoming data by key columns
main_title(lsting) <- "this is some title"
main_footer(lsting) <- "this is some footer"
cat(export_as_txt(lsting, file = NULL, paginate = TRUE))
#> c("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c("AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345", "AB12345")\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c("id-134", "id-134", "id-134", "id-134", "id-141", "id-141", "id-141", "id-141", "id-141", "id-141", "id-236", "id-236", "id-236", "id-265", "id-265", "id-265", "id-265", "id-42", "id-42", "id-42", "id-42", "id-42", "id-42", "id-42", "id-42")\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c("BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1"), c(47, 47, 47, 47, 35, 35, 35, 35, 35, 35, 32, 32, 32, 25, 25, 25, 25, 36, 36, 36, 36, 36, 36, 36, 36), c(2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 2, 2, 2, 3, 3, 3, 3, 2, 2, 2, 2, 2, 2, 2, 2), c(3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c("BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1", "BRA-1")\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 3, 3, 3, 3, 3, 3, 2, 2, 2, 3, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1), c(1, 1, 1, 1, 3, 3, 3, 3, 3, 3, 2, 2, 2, 3, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 3, 3, 3, 3, 3, 3, 2, 2, 2, 3, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1), c(1, 1, 1, 1, 3, 3, 3, 3, 3, 3, 2, 2, 2, 3, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 1)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(6.46299057842479, 6.46299057842479, 6.46299057842479, 6.46299057842479, 7.51607612428241, 7.51607612428241, 7.51607612428241, 7.51607612428241, 7.51607612428241, 7.51607612428241, 7.66300121077566, 7.66300121077566, 7.66300121077566, 10.323346349886, 10.323346349886, 10.323346349886, 10.323346349886, 2.26753940777848, 2.26753940777848, 2.26753940777848, 2.26753940777848, 2.26753940777848, 2.26753940777848, 2.26753940777848, 2.26753940777848)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 3, 3, 3, 3, 3, 3, 3, 3, 3, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 1)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(18787, 18787, 18787, 18787, 18683, 18683, 18683, 18683, 18683, 18683, 18856, 18856, 18856, 18391, 18391, 18391, 18391, 18480, 18480, 18480, 18480, 18480, 18480, 18480, 18480)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1623331613.9562, 1623331613.9562, 1623331613.9562, 1623331613.9562, 1614556036.9562, 1614556036.9562, 1614556036.9562, 1614556036.9562, 1614556036.9562, 1614556036.9562, 1629569605.9562, 1629569605.9562, 1629569605.9562, 1589330292.9562, 1589330292.9562, 1589330292.9562, 1589330292.9562, 1596782699.9562, 1596782699.9562, 1596782699.9562, 1596782699.9562, 1596782699.9562, 1596782699.9562, 1596782699.9562, 1596782699.9562)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1686445517.9562, 1686445517.9562, 1686445517.9562, 1686445517.9562, 1677669940.9562, 1677669940.9562, 1677669940.9562, 1677669940.9562, 1677669940.9562, 1677669940.9562, 1692683509.9562, 1692683509.9562, 1692683509.9562, 1631978615.9562, 1631978615.9562, 1631978615.9562, 1631978615.9562, NA, NA, NA, NA, NA, NA, NA, NA)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3), c(19519, 19519, 19519, 19519, 19417, 19417, 19417, 19417, 19417, 19417, 19591, 19591, 19591, 18888, 18888, 18888, 18888, NA, NA, NA, NA, NA, NA, NA, NA)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(731, 731, 731, 731, 731, 731, 731, 731, 731, 731, 731, 731, 731, 494, 494, 494, 494, NA, NA, NA, NA, NA, NA, NA, NA)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 5, 5, 5, 5, NA, NA, NA, NA, NA, NA, NA, NA)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(19519, 19519, 19519, 19519, 19417, 19417, 19417, 19417, 19417, 19417, 19591, 19591, 19591, 18888, 18888, 18888, 18888, NA, NA, NA, NA, NA, NA, NA, NA), c(19537, 19537, 19537, 19537, 19446, 19446, 19446, 19446, 19446, 19446, 19614, 19614, 19614, 18908, 18908, 18908, 18908, NA, NA, NA, NA, NA, NA, NA, NA)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904, 63113904)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(3, 4, 1, 2, 3, 4, 5, 1, 2, 6, 1, 2, 3, 1, 4, 2, 3, 4, 6, 8, 5, 7, 1, 3, 2)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(3, 4, 1, 2, 3, 4, 5, 1, 2, 6, 1, 2, 3, 1, 4, 2, 3, 4, 6, 8, 5, 7, 1, 3, 2)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(2, 2, 4, 9, 1, 2, 1, 4, 10, 8, 3, 3, 3, 7, 6, 9, 8, 2, 2, 2, 5, 3, 7, 7, 8)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(2, 2, 4, 9, 1, 2, 1, 4, 10, 8, 3, 3, 3, 7, 6, 9, 8, 2, 2, 2, 5, 3, 7, 7, 8), c(2, 2, 4, 9, 1, 2, 1, 4, 10, 8, 3, 3, 3, 7, 6, 9, 8, 2, 2, 2, 5, 3, 7, 7, 8)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 3, 8, 1, 1, 1, 3, 9, 7, 2, 2, 2, 6, 5, 8, 7, 1, 1, 1, 4, 2, 6, 6, 7), c(1, 1, 3, 7, 1, 1, 1, 3, 8, 7, 2, 2, 2, 6, 5, 7, 7, 1, 1, 1, 4, 2, 6, 6, 7)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 3, 6, 1, 1, 1, 3, 7, 6, 2, 2, 2, 5, 4, 6, 6, 1, 1, 1, 3, 2, 5, 5, 6)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(2, 2, 2, 2, 1, 2, 1, 2, 1, 3, 3, 3, 3, 2, 3, 2, 3, 2, 2, 2, 1, 3, 2, 2, 3), c(2, 2, 1, 1, 1, 2, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 1, 1, 1, 2)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 2, 2, 2, 1), c(1666224000, 1675814400, 1644969600, 1649548800, 1657065600, 1666310400, 1669334400, 1636848000, 1640649600, 1677542400, 1631404800, 1645142400, 1690156800, 1593475200, 1624147200, 1606608000, 1623196800, 1610582400, 1636502400, 1637884800, 1621987200, 1637366400, 1600128000, 1604966400, 1602374400)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(1685923200, 1681516800, 1668038400, 1671580800, 1659052800, 1674345600, 1673049600, 1637452800, 1676937600, 1677628800, 1666828800, 1689811200, 1690502400, 1597276800, 1627344000, 1620604800, 1625011200, 1632960000, 1653264000, 1647561600, 1626307200, 1648598400, 1643846400, 1631404800, 1656720000)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(497, 608, 251, 304, 493, 600, 635, 259, 303, 730, 22, 181, 702, 48, 403, 200, 392, 160, 460, 476, 292, 470, 39, 95, 65)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(725, 674, 518, 559, 516, 693, 678, 266, 723, 731, 432, 698, 706, 92, 440, 362, 413, 419, 654, 588, 342, 600, 545, 401, 694)\s\nc("AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-134", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-141", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-236", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-265", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", 
#> "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42", "AB12345-BRA-1-id-42"), c("cl A", "cl A", "cl B", "cl D", "cl A", "cl A", "cl A", "cl B", "cl D", "cl D", "cl B", "cl B", "cl B", "cl C", "cl C", "cl D", "cl D", "cl A", "cl A", "cl A", "cl B", "cl B", "cl C", "cl C", "cl D"), c(2, 2, 3, 3, 1, 2, 1, 3, 1, 5, 5, 5, 5, 2, 4, 3, 5, 2, 2, 2, 1, 5, 2, 2, 5), c("Y / N", "Y / N", "N / N", "N / N", "N / N", "Y / N", "N / N", "N / N", "N / Y", "Y / N", "N / Y", "N / Y", "N / Y", "N / Y", "N / Y", "N / N", "Y / N", "Y / N", "Y / N", "Y / N", "Y / N", "N / Y", "N / Y", "N / Y", "Y / N")