Create a Listing from a data.frame or tibble

Create a Listing from a data.frame or tibble

Usage

as_listing(
  df,
  cols = key_cols,
  key_cols = names(df)[1],
  main_title = NULL,
  subtitles = NULL,
  main_footer = NULL,
  prov_footer = NULL
)

as_keycol(vec)

is_keycol(vec)

get_keycols(df)

# S4 method for listing_df
matrix_form(obj, indent_rownames = FALSE)

listing_dispcols(df)

add_listing_dispcol(df, new)

listing_dispcols(df) <- value

add_listing_col(df, name, fun = NULL, format = NULL)

Arguments

df

listing_df. The listing to modify.

cols

character. Names of columns (including but not limited to key columns) which should be displayed when the listing is rendered.

key_cols

character. Names of columns which should be treated as *key columns* when rendering the listing.

main_title

character(1) or NULL. The main title for the listing, or `NULL` (the default). Must be length 1 non-NULL.

subtitles

character or NULL. A vector of subtitle(s) for the listing, or NULL (the default).

main_footer

character or NULL. A vector of main footer lines for the listing, or `NULL` (the default).

prov_footer

character or NULL. A vector of provenance strings for the listing, or `NULL` (the default).

vec

any. A column vector from a `listing_df`

obj

ANY. Object to be transformed into a ready-to-render form (a MatrixPrintForm object)

indent_rownames

logical(1), if TRUE the column with the row names in the `strings` matrix of has indented row names (strings pre-fixed)

new

character. Names of columns to be added to the set of display columns.

value

character. New value.

name

character(1). Name of the existing or new column to be displayed when the listing is rendered

fun

function or NULL. A function which accepts df and returns the vector for a new column, which is added to df as name, or NULL if marking an existing column as a listing column

format

FormatSpec. A format specification (format string, function, or sprintf format) for use when displaying the column during rendering.

Value

A `listing_df` object, sorted by the key columns.

`df`, with `name` created (if necessary) and marked for display during rendering.

Examples

dat <- ex_adae

lsting <- as_listing(dat[1:25,], key_cols = c("USUBJID", "AESOC")) %>%
    add_listing_col("AETOXGR") %>%
    add_listing_col("BMRKR1", format = "xx.x") %>%
    add_listing_col("AESER / AREL", fun = function(df) paste(df$AESER, df$AREL, sep = " / "))
#> sorting incoming data by key columns

mat <- matrix_form(lsting)

cat(toString(mat))
#> Unique Subject Identifier   Primary System Organ Class   Analysis Toxicity Grade   Continous Level Biomarker 1   AESER / AREL
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> AB12345-BRA-1-id-134        cl A                                    2                          6.5                  Y / N    
#>                                                                     2                          6.5                  Y / N    
#>                             cl B                                    3                          6.5                  N / N    
#>                             cl D                                    3                          6.5                  N / N    
#> AB12345-BRA-1-id-141        cl A                                    1                          7.5                  N / N    
#>                                                                     2                          7.5                  Y / N    
#>                                                                     1                          7.5                  N / N    
#>                             cl B                                    3                          7.5                  N / N    
#>                             cl D                                    1                          7.5                  N / Y    
#>                                                                     5                          7.5                  Y / N    
#> AB12345-BRA-1-id-236        cl B                                    5                          7.7                  N / Y    
#>                                                                     5                          7.7                  N / Y    
#>                                                                     5                          7.7                  N / Y    
#> AB12345-BRA-1-id-265        cl C                                    2                         10.3                  N / Y    
#>                                                                     4                         10.3                  N / Y    
#>                             cl D                                    3                         10.3                  N / N    
#>                                                                     5                         10.3                  Y / N    
#> AB12345-BRA-1-id-42         cl A                                    2                          2.3                  Y / N    
#>                                                                     2                          2.3                  Y / N    
#>                                                                     2                          2.3                  Y / N    
#>                             cl B                                    1                          2.3                  Y / N    
#>                                                                     5                          2.3                  N / Y    
#>                             cl C                                    2                          2.3                  N / Y    
#>                                                                     2                          2.3                  N / Y    
#>                             cl D                                    5                          2.3                  Y / N