toString method for listing_df object

toString method for listing_df object

Usage

# S4 method for listing_df
toString(x, widths = NULL, col_gap = 3, hsep = default_hsep())

Arguments

x

listing_df object

widths

numeric (or NULL). (proposed) widths for the columns of x. The expected length of this numeric vector can be retrieved with ncol() + 1 as the column of row names must also be considered.

col_gap

numeric(1). Space (in characters) between columns

hsep

character(1). Characters to repeat to create header/body separator line.

Value

A character value containing the listing rendered into ASCII text.

Examples

lsting <- as_listing(mtcars)
#> sorting incoming data by key columns
toString(lsting)
#> [1] "———————————————————————————————————————————————————————————————————————\nmpg    cyl   disp    hp    drat    wt     qsec    vs   am   gear   carb\n———————————————————————————————————————————————————————————————————————\n10.4    8     472    205   2.93   5.25    17.98   0    0     3      4  \n        8     460    215    3     5.424   17.82   0    0     3      4  \n13.3    8     350    245   3.73   3.84    15.41   0    0     3      4  \n14.3    8     360    245   3.21   3.57    15.84   0    0     3      4  \n14.7    8     440    230   3.23   5.345   17.42   0    0     3      4  \n15      8     301    335   3.54   3.57    14.6    0    1     5      8  \n15.2    8    275.8   180   3.07   3.78     18     0    0     3      3  \n        8     304    150   3.15   3.435   17.3    0    0     3      2  \n15.5    8     318    150   2.76   3.52    16.87   0    0     3      2  \n15.8    8     351    264   4.22   3.17    14.5    0    1     5      4  \n16.4    8    275.8   180   3.07   4.07    17.4    0    0     3      3  \n17.3    8    275.8   180   3.07   3.73    17.6    0    0     3      3  \n17.8    6    167.6   123   3.92   3.44    18.9    1    0     4      4  \n18.1    6     225    105   2.76   3.46    20.22   1    0     3      1  \n18.7    8     360    175   3.15   3.44    17.02   0    0     3      2  \n19.2    6    167.6   123   3.92   3.44    18.3    1    0     4      4  \n        8     400    175   3.08   3.845   17.05   0    0     3      2  \n19.7    6     145    175   3.62   2.77    15.5    0    1     5      6  \n21      6     160    110   3.9    2.62    16.46   0    1     4      4  \n        6     160    110   3.9    2.875   17.02   0    1     4      4  \n21.4    6     258    110   3.08   3.215   19.44   1    0     3      1  \n        4     121    109   4.11   2.78    18.6    1    1     4      2  \n21.5    4    120.1   97    3.7    2.465   20.01   1    0     3      1  \n22.8    4     108    93    3.85   2.32    18.61   1    1     4      1  \n        4    140.8   95    3.92   3.15    22.9    1    0     4      2  \n24.4    4    146.7   62    3.69   3.19     20     1    0     4      2  \n26      4    120.3   91    4.43   2.14    16.7    0    1     5      2  \n27.3    4     79     66    4.08   1.935   18.9    1    1     4      1  \n30.4    4    75.7    52    4.93   1.615   18.52   1    1     4      2  \n        4    95.1    113   3.77   1.513   16.9    1    1     5      2  \n32.4    4    78.7    66    4.08    2.2    19.47   1    1     4      1  \n33.9    4    71.1    65    4.22   1.835   19.9    1    1     4      1  \n\n———————————————————————————————————————————————————————————————————————\n"