Methods for listing_df objects

See core documentation in formatters for descriptions of these functions.

Usage

# S3 method for listing_df
print(x, widths = NULL, tf_wrap = FALSE, max_width = NULL, ...)

# S4 method for listing_df
toString(x, ...)

# S4 method for listing_df
[(x, i, j, drop = FALSE)

# S4 method for listing_df
main_title(obj)

# S4 method for listing_df
subtitles(obj)

# S4 method for listing_df
main_footer(obj)

# S4 method for listing_df
prov_footer(obj)

# S4 method for listing_df
main_title(obj) <- value

# S4 method for listing_df
subtitles(obj) <- value

# S4 method for listing_df
main_footer(obj) <- value

# S4 method for listing_df
prov_footer(obj) <- value

# S4 method for listing_df
num_rep_cols(obj)

Arguments

x

listing_df. The listing.

widths

numeric (or NULL). (proposed) widths for the columns of x. The expected length of this numeric vector can be retrieved with ncol() + 1 as the column of row names must also be considered.

tf_wrap

logical(1). Should the texts for title, subtitle, and footnotes be wrapped?

max_width

integer(1), character(1) or NULL. Width that title and footer (including footnotes) materials should be word-wrapped to. If NULL, it is set to the current print width of the session (getOption("width")). If set to "auto", the width of the table (plus any table inset) is used. Ignored completely if tf_wrap is FALSE.

...

dots. See toString method in formatters for all parameters.

i

ANY. Passed to base [ methods.

j

ANY. Passed to base [ methods.

drop

For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.

obj

The object.

value

typically an array-like R object of a similar class as x.

Value

for getter methods, the value of the aspect of obj; for setter methods, obj with the relevant element of the listing updated.

Examples

lsting <- as_listing(mtcars)
#> sorting incoming data by key columns
main_title(lsting) <- "Hi there"

main_title(lsting)
#> [1] "Hi there"